+++ examples/roles/dummy/tasks/main.rb
@@ -1,7 +1,27 @@
 task 'say hello' do
-  result = command jinja('a_command | default("ls howdy")') do
-    chdir '/tmp'
-  end
+  with_items(%w(a b c)) do |item|
+    result = shell "ls #{item} || true" do
+      chdir '/tmp'
+    end
 
-  changed_when "'howdy' in #{result.stdout}"
+    failed_when "'No such filesss' in #{result.stderr}"
+  end
+end
+
+task 'and goodbye' do
+  stuff = {
+    cmd1: {
+      foobar: '123'
+    },
+    cmd2: {
+      foobar: '456'
+    }
+  }
+
+  with_dict(stuff) do |key, value|
+    # will run ls cmd1 123 and ls cmd2 456
+    command "ls #{key} #{value.foobar}" do
+      chdir '/tmp'
+    end
+  end
 end