+++ lib/ansible/ruby/models/task.rb
@@ -19,7 +19,9 @@ module Ansible
         attribute :when
         validates :when, type: String
         attribute :with_dict
-        validates :with_dict, type: String
+        validates :with_dict, type: MultipleTypes.new(String, Hash)
+        attribute :with_items
+        validates :with_items, type: MultipleTypes.new(String, Array)
         attribute :notify
         validates :notify, type: TypeGeneric.new(String)
         attribute :async
@@ -28,6 +30,7 @@ module Ansible
         validates :poll, type: Integer
         attribute :ignore_errors
         validates :ignore_errors, type: MultipleTypes.new(TrueClass, FalseClass)
+        validate :loop_and_dict
 
         def to_h
           result = super
@@ -44,6 +47,12 @@ module Ansible
           end
           new_result[:notify] = [*notify] if notify
           new_result
+        end
+
+        private
+
+        def loop_and_dict
+          errors.add :with_items, 'Cannot use both with_items and with_dict!' if with_items && with_dict
         end
       end
     end