+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Under development
 * Add jinja variable helper
+* Add with_items support and add block helper for that and with_dict
 
 # 1.0.2 - 25 August 2016
 * Allow free form tasks to be supplied without a block